Add files via upload
Ticket ID A-2703 raised concerns about a suspicious email received by an employee. The email claimed to be from "Def Communications" but had inconsistencies in the sender's details and grammatical errors. It included an attachment named "bfsvc.exe," which was found to be a known harmful file. The email was flagged as potentially dangerous because it could lead to malware being downloaded onto the company's systems. As a precaution, the issue was escalated to a higher level for further investigation and action.